+++ tempest_fastapi_sdk/__init__.py
@@ -242,7 +242,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.36.0"
+__version__: str = "0.37.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",